refactor(ingest): extract signal evidence and edge types into signals/types
Moves RecordRefLike, JsonRecord, ToolCallLike, TurnRow, SessionTurns,
CorrectionEdge, ProposedEdge, SkillPairAccum, RecoveryEdge,
DerivedFrictionEvent, DerivedDiagnosticEvent verbatim into
signals/types.ts and adds the SignalEvidence/DerivedSignals aggregates
the core will consume. derive-signals.ts re-imports them; a transitional
re-export keeps signals/core.test.ts compiling until Task 5 prunes it.
